requirements/base.txt
six==1.10.0
packaging==16.8
appdirs==1.4.3
contextlib2==0.5.4
coreapi==2.3.0
coreschema==0.0.4
Django==1.11
django-countries==4.4
django-model-fields==0.1.0
django-rest-swagger==2.1.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
django-environ==0.4.1
django-authtools==1.5.0
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
psycopg2==2.7.1
openapi-codec==1.3.1
PyJWT==1.4.2
pyparsing==2.2.0
pytz==2017.2
raven==6.0.0
requests==2.13.0
simplejson==3.10.0
sqlparse==0.2.3
uritemplate==3.0.0
django_cron==0.5.0
django-pyowm==0.2.0
